Rename WalletApi file
