+++ src/api/wallet/WalletApiMock.ts
@@ -4,7 +4,7 @@ import assert from 'assert'
 
 import { log, wait, toWei } from 'utils'
 import { USER_1, USER_2 } from '../../../test/data'
-import { WalletApi, WalletInfo, ProviderInfo } from './WalletApiImpl'
+import { WalletApi, WalletInfo, ProviderInfo } from './WalletApi'
 
 type OnChangeWalletInfo = (walletInfo: WalletInfo) => void
