+++ src/hooks/useWalletConnection.ts
@@ -2,7 +2,7 @@ import { walletApi } from 'api'
 import { useEffect } from 'react'
 import { Command } from 'types'
 import useSafeState from './useSafeState'
-import { WalletInfo } from 'api/wallet/WalletApiImpl'
+import { WalletInfo } from 'api/wallet/WalletApi'
 
 export const useWalletConnection = (): WalletInfo => {
   const [walletInfo, setWalletInfo] = useSafeState<WalletInfo>({ isConnected: false })
