+++ test/components/EtherscanLink.components.test.tsx
@@ -5,7 +5,7 @@ import { EtherscanLink } from 'components/EtherscanLink'
 import { TX_HASH, USER_1, CONTRACT, TOKEN_1 } from '../data'
 import { abbreviateString } from 'utils'
 import { Network } from 'types'
-import { WalletInfo } from 'api/wallet/WalletApiImpl'
+import { WalletInfo } from 'api/wallet/WalletApi'
 
 let isConnected: boolean
 let networkId: Network | undefined
